query1 data type migrations + fixes

+++ tpe2-g7/client/src/main/java/ar/edu/itba/pod/hazelcast/client/ClientQuery3.java
@@ -71,9 +71,9 @@ public class ClientQuery3 {
         for (String[] bike : bikesCSV) {
             bikes.add(new Trip(
                     LocalDateTime.parse(bike[0], DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")),
-                    Integer.parseInt(bike[1]),
+                    Long.parseLong(bike[1]),
                     LocalDateTime.parse(bike[2], DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")),
-                    Integer.parseInt(bike[3]),
+                    Long.parseLong(bike[3]),
                     Boolean.parseBoolean(bike[4])
             ));
         }
